Flat and Tree Traits
